Move authorities to constants folder
